[Refactor] PointHisotry -> User 통합
