+++ src/main/java/com/inq/wishhair/wesharewishhair/user/controller/dto/response/MyPageResponse.java
@@ -1,7 +1,7 @@
 package com.inq.wishhair.wesharewishhair.user.controller.dto.response;
 
 import com.inq.wishhair.wesharewishhair.domain.login.dto.UserSessionDto;
-import com.inq.wishhair.wesharewishhair.point.domain.PointHistory;
+import com.inq.wishhair.wesharewishhair.user.domain.point.domain.PointHistory;
 import com.inq.wishhair.wesharewishhair.user.domain.User;
 import com.inq.wishhair.wesharewishhair.user.enums.Sex;
 import lombok.AllArgsConstructor;
