+++ src/main/java/com/inq/wishhair/wesharewishhair/user/domain/point/domain/PointHistory.java
@@ -1,7 +1,7 @@
-package com.inq.wishhair.wesharewishhair.point.domain;
+package com.inq.wishhair.wesharewishhair.user.domain.point.domain;
 
 import com.inq.wishhair.wesharewishhair.auditing.BaseEntity;
-import com.inq.wishhair.wesharewishhair.point.enums.PointType;
+import com.inq.wishhair.wesharewishhair.user.domain.point.enums.PointType;
 import com.inq.wishhair.wesharewishhair.user.domain.User;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
