+++ src/main/java/com/inq/wishhair/wesharewishhair/user/domain/point/domain/PointRepository.java
@@ -1,6 +1,5 @@
-package com.inq.wishhair.wesharewishhair.point.domain;
+package com.inq.wishhair.wesharewishhair.user.domain.point.domain;
 
-import com.inq.wishhair.wesharewishhair.point.domain.PointHistory;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
@@ -8,7 +7,7 @@ import org.springframework.data.repository.query.Param;
 
 import java.util.List;
 
-public interface PointHistoryRepository extends JpaRepository<PointHistory, Long> {
+public interface PointRepository extends JpaRepository<PointHistory, Long> {
 
     @Query("select p from PointHistory p " +
             "where p.user.id = :userId " +
